drm/msm/mdp5: Ensure clocks are enabled in legacy cursor disable path
When userspace uses the legacy DRM_IOCTL_MODE_CURSOR ioctl to disable
the cursor, we can end up in a situation where we try to configure
registers without the MDP clocks being on. Unlike other legacy ABI,
this ioctl doesn't use atomic, resulting in the prepare_commit()
kms func to be called before we commit to the new atomic state.
Make sure clocks are on in the cursor disable path.

(Can we end up in a similar situation if userspace decides to call
this ioctl to enable cursor, before enabling the CRTC?)

This isn't the nicest solution, need to dig in a bit more to figure
out something better.

Signed-off-by: Archit Taneja <architt@codeaurora.org>
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>
Signed-off-by: Ilia Lin <ilialin@codeaurora.org>
(cherry picked from commit 91eecf302625ad32bb0a836d287282b803e61b19)