drm/bridge: adv7511: Power on the chip when cable is plugged back
When a HPD unplug occurs, the chip resets itself. When the connector
is plugged in again, we need to restore its state (i.e, power it on)
for it to be configured properly.

This was done correctly before, but the commit 01d847af1c04
"drm/bridge: adv7511: Switch to using drm_kms_helper_hotplug_event()"
changes what we do when we get a HPD interrupt. Previously, the
drm_helper_hpd_irq_event() internally called the connector's detect() op
(adv7511_detect), which powered on the chip if this was a
unplugged -> plugged transition. After the patch, the interrupt handler
schedules a work that only reports the connection status.

This fixes a regression where HDMI isn't powered on once unplugged and
plugged in again.

Reported-by: Ilia Lin <ilial@codeaurora.org>
Signed-off-by: Archit Taneja <architt@codeaurora.org>
(cherry picked from commit 7a1690cbc7633f49cdb976687d32d5d2affbe14f)