drm/bridge/adv7511: Always poll for EDID_READY
With the current code, if we have a irq handler registered, the driver
tries to wait on the EDID_READY interrupt using
wait_event_interruptible_timeout.

With the following sequence (with OE rootfs), we don't get a EDID ready
interrupt:

1. Boot up with cable unplugged.
2. Run X.
3. Insert cable.
4. Run xrandr.

In the above sequence, when we insert the cable(step 3), we get a hot plug
interrupt, followed by the driver trying to read the EDID for the first
time. This EDID read does result in a EDID_READY interrupt being
triggered, and the EDID is read successfully. However, when we run xrandr
(step 4), the driver tries to read the EDID again, and this time the
EDID_READY interrupt isn't triggered.

One experiment done was to insert a delay between the __adv7511_power_on
and drm_do_get_edid() in adv7511_get_modes(). This resulted in the
interrupt firing, but the delay required to get the interrupt varied
across different boards.

Instead of relying on EDID_READY interrupt, we poll on the interrupt
status register (ADV7511_REG_INT(0)) to wait for the EDID_READY field to
be set.

The reason why the EDID_READY interrupt doesn't fire only for the above
sequence hasn't been root-caused. We may want to consider switching back
to using interrupts for EDID_READY once it is root-caused.

Signed-off-by: Archit Taneja <architt@codeaurora.org>